refactor: centralize status mapping and improve toast management in applicant components
